+++ src/core/errors/custom-errors/custom.error.ts
