+++ src/core/errors/custom-errors/internal-server.error.ts
@@ -1,4 +1,4 @@
-import { CustomError } from "./customError.js";
+import { CustomError } from "./custom.error.js";
 
 export class InternalServerError extends CustomError {
   constructor({
