+++ src/core/errors/custom-errors/missing-secret-key.error.ts
@@ -1,4 +1,4 @@
-import { CustomError } from "./customError.js";
+import { CustomError } from "./custom.error.js";
 
 export class MissingSecretKeyError extends CustomError {
   constructor({
