--- src/core/errors/importAllErrors.ts
@@ -1,11 +1,0 @@
-export * from "./customErrors/badInputError.js";
-export * from "./customErrors/customError.js";
-export * from "./customErrors/dateBirthdayFutureError.js";
-export * from "./customErrors/userAlreadyExistsError.js";
-export * from "./customErrors/unauthorizedUser.js";
-export * from "./customErrors/notFoundError.js";
-export * from "./customErrors/missingSecretKeyError.js";
-export * from "./customErrors/missingCredentialsAdminError.js";
-export * from "./customErrors/maximumAgeError.js";
-export * from "./customErrors/invalidDateFormatError.js";
-export * from "./customErrors/internalServerError.js";
