+++ src/core/utils/hash-utils.ts
